Add WrapFilter and implement AuthFilter with it
This commit introduces the notion of a "WrapFilter" which wraps a
(T, Req) => Resp service with a Req => Resp service using a lookup
function.

The most obvious use case for this is the AuthFilter. The lookup filter
looks at some cookie id, and returns the associated user, passing it to
the internal service. If there is no user, it returns Future.never (a
value that can be configured)

+++ src/test/scala/com/leeavital/unfangled/FilterSpec.scala
@@ -17,7 +17,7 @@ class FilterSpec extends FlatSpec with Matchers {
   }
 
   // this turns a Double => String into an Int => Float
-  object IntDoubler extends Filter[Int, String, Double, Float] {
+  object IntDoubler extends Filter[Double, String, Int, Float] {
     def apply(req: Int, service: PartialFunction[Double, Future[String]]): Future[Float] = {
       val d = req.toDouble
       service(d).map(x => x.toDouble.toFloat * 2)
